+++ CHANGELOG.md
@@ -10,7 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - Show correct placeholder in legacy consumer in input field template. (#5924)
-- Do not allow choosing placeholder in legacy consumer in select field template. (#5924)
+- Do not allow to choosing placeholder in legacy consumer in select field template if field is required. (#5924)
 - Remove `give_` prefix from field names in legacy consumer in select, radio and textarea field template. (#5924)
 - Set select field name to array type in legacy consumer for select field template if multiple value acceptable field template. (#5924)
 - Sanitize array type values correctly before make custom field value persistence. (#5924)