+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+## 2.13.2 - 2021-08-26
+
 ## Fixed
 
 - Admin able to create donation form and edit existing donation form if stripe disabled. (#5935)